[AND-113] Move and reuse method to get installed app icons
